requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
django-import-export==2.8.0
Django==4.1
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pip==22.2.2
pydantic==1.9.1
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
setuptools==60.2.0
sqlparse==0.4.2
tablib==3.2.1
typing-extensions==4.3.0
tzdata==2022.1
wheel==0.37.1
xlrd==2.0.1
xlwt==1.3.0
